Change to only teardown models with fixtures
When tearDownFixtures is called, only models where
there is a corresponding fixture will be torn down.

I would consider this as a breaking change and we should
bump the major version.

It may be sensible to add some kind of configuration
around this so that the behavior can be enabled and
disabled