Fail silently by default to maintain backwards compatibility
v1.0 and earlier would not report errors if fixture setup failed.
This may be expected behavior and/or tested behavior in clients'
applications, so switching to show errors by default is a breaking
change.

This version reverts to the earlier silent-by-default behavior and
adds a component option (`errorOnSetupFailure`) to enable it.
The component-powered version will also throw an error if the server
is supposed to load fixtures on startup but cannot.

Also adds debug messages and Istanbul test coverage.